+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   inserts relative paths.
 - More reactive updating of the related files section
 - Improvements during insertion of snippets
+- The footnote editor now has a consistent background color in dark mode
 
 ## Under the Hood